Drop CircleCI, move to Mega Linter with code from template
This replaces the cibuild that just runs terraform fmt.

Linting didn't like trying to set a default paramater file, so dropped.
Ideally in the future we can use some terraform to set it to null, by default.

Also increased the default version to a more modern version and synced
between variables and readme

Added gitignore file as well.